Rename model package to spec
