+++ internal/spec/build_test.go
@@ -1,4 +1,4 @@
-package model
+package spec
 
 import (
 	"runtime"
